pyproject.toml
[project]
name = "llamaindex-demo"
version = "0.1.0"
description = "基于LlamaIndex + LangChain的教育行业RAG智能教案生成系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LlamaIndex 核心包
    "llama-index>=0.9.48",
    "llama-index-llms-openai>=0.1.13",
    "llama-index-embeddings-openai>=0.1.9",
    "llama-index-vector-stores-chroma>=0.1.7",
    
    # LangChain 核心包
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langchain-experimental>=0.0.47",
    "langsmith>=0.0.83",
    
    # LangChain 文档加载器
    "langchain-text-splitters>=0.0.1",
    "unstructured>=0.11.8",
    "unstructured[pdf]>=0.11.8",
    "unstructured[docx]>=0.11.8",
    
    # Web框架
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    
    # 数据处理
    "pandas>=2.1.4",
    "numpy>=1.25.2",
    "python-docx>=1.1.0",
    "pypdf2>=3.0.1",
    
    # 向量数据库
    "chromadb>=0.4.15",
    "faiss-cpu>=1.7.0",
    
    # 环境配置
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    
    # HTTP客户端
    "httpx>=0.25.2",
    "aiohttp>=3.9.1",
    
    # 可视化和工具
    "streamlit>=1.28.2",
    "plotly>=5.17.0",
    
    # 测试
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    
    # 额外工具
    "tiktoken>=0.5.0",
    "openai>=1.3.0",
]
